Implement OAuth2 authorization security measures
The OAuth2 authorization flow now has a solid security implementation.
When the user requests an authorization link from the server, two nonces
are created. One is associated to the user through a secure cookie, and
the other is put into the OAuth2 authorization URL as the "state"
parameter. Once the user authorizes the application, they're redirected
to the token-fetching endpoint, which checks in the server's database
whether the two nonces match. If not, the endpoint will not proceed with
getting tokens, and the user will not be authorized to use the bot.